The ZAMQP_LOGLEVEL environment variable was not yet accessible at import time. Moved Logger to utils module to delay the import.
