+++ src/collective/zamqp/keepalive.py
@@ -53,7 +53,7 @@ from grokcore import component as grok
 
 from zope.component import getUtility
 
-from collective.zamqp import logger
+from collective.zamqp.utils import logger
 from collective.zamqp.producer import Producer
 from collective.zamqp.consumer import Consumer
 from collective.zamqp.interfaces import IProducer
