+++ src/collective/zamqp/utils.py
@@ -8,6 +8,8 @@
 ###
 """Helpers for re-using amqp-supporting packages between different buildouts"""
 
+import logging
+import os
 from App.config import getConfiguration
 
 
@@ -21,3 +23,21 @@ def getBuildoutName():
     else:
         # configuration not found, return some sane default string
         return 'collective.zamqp.default'
+
+
+class Logger(object):
+
+    def __init__(self, name, default_level):
+        self.logger = logging.getLogger(name)
+        self.default_level = default_level
+
+    def default(self, message, *args, **kwargs):
+        self.logger.log(self.default_level, message, *args, **kwargs)
+
+    def __getattr__(self, item):
+        return getattr(self.logger, item)
+
+logger = Logger('collective.zamqp', default_level={
+    'DEBUG': logging.DEBUG,
+    'INFO': logging.INFO,
+}.get(os.environ.get('ZAMQP_LOGLEVEL'), logging.DEBUG))
